requirements.txt
fastapi
gunicorn
uvicorn
jinja2  # for templates
pytest
requests
httpx
pre-commit # for pre-commit hooks(steps), automatically run on commit
python-dotenv # for loading environment variables from .env file
python-multipart
pillow
aiofiles
pytesseract